Implemented: - New class Scene for storing and handling all scene objects. - Material components for objects.
Changed:
- Restructured the handling of scene objects in class Camera for a more general approach.

+++ RayTracing/Ray.h
@@ -4,8 +4,8 @@
 
 struct Ray
 {
-	glm::vec3 start_point;
-	glm::vec3 direction;
+	glm::dvec3 start_point;
+	glm::dvec3 direction;
 
-	inline Ray(const glm::vec3& start, const glm::vec3& direction) : start_point(start), direction(direction) { }
+	inline Ray(const glm::dvec3& start, const glm::dvec3& direction) : start_point(start), direction(direction) { }
 };
